+if body exists

+++ infty/core/signals.py
@@ -16,54 +16,56 @@ def execute_after_save(sender, instance, created, *args, **kwargs):
 @receiver(models.signals.post_save, sender=Topic)
 def topic_post_save(sender, instance, created, *args, **kwargs):
 
-    html = mistune.markdown(instance.body)
+    if instance.body:
 
-    # Parse Needs
-    soup = bs4.BeautifulSoup(
-        mistune.markdown(instance.body),
-        'html.parser'
-    )
+        html = mistune.markdown(instance.body)
 
-    need_lists = soup.find_all(
-        'code', {'class': 'lang-yml'}
-    )
+        # Parse Needs
+        soup = bs4.BeautifulSoup(
+            mistune.markdown(instance.body),
+            'html.parser'
+        )
 
-    needs = []
-    for need_list in need_lists:
-        blocks = yaml.load(need_list.text)
-        for block in blocks:
-            if 'needs' in block.keys():
-                needs.extend(block['needs'])
+        need_lists = soup.find_all(
+            'code', {'class': 'lang-yml'}
+        )
 
-    # Retrieve Existing Needs:
-    Needs = instance.parents.filter(type=0)
+        needs = []
+        for need_list in need_lists:
+            blocks = yaml.load(need_list.text)
+            for block in blocks:
+                if 'needs' in block.keys():
+                    needs.extend(block['needs'])
 
-    lang = instance.languages[0] if instance.languages else 'en'
+        # Retrieve Existing Needs:
+        Needs = instance.parents.filter(type=0)
 
-    # Create and Link Needs
-    for need in needs:
+        lang = instance.languages[0] if instance.languages else 'en'
 
-        title = '.:{}:{}'.format(lang, need.get('name'))
-        tool = [key for key in need.keys() if key not in ['name']][0]
-        body = '.:{}\n```{}: {}```'.format(lang, tool, need.get(tool))
+        # Create and Link Needs
+        for need in needs:
 
-        if not Needs.filter(body=body).exists():
+            title = '.:{}:{}'.format(lang, need.get('name'))
+            tool = [key for key in need.keys() if key not in ['name']][0]
+            body = '.:{}\n```{}: {}```'.format(lang, tool, need.get(tool))
 
-            topic = Topic.objects.create(
-                title=title,
-                body=body,
-                owner=instance.owner,
-                type=0
-            )
+            if not Needs.filter(body=body).exists():
 
-            instance.parents.add(topic)
+                topic = Topic.objects.create(
+                    title=title,
+                    body=body,
+                    owner=instance.owner,
+                    type=0
+                )
 
-    # Remove needs that are not
+                instance.parents.add(topic)
 
-    tool = lambda _: [key for key in _.keys() if key not in ['name']][0]
+        # Remove needs that are not
+
+        tool = lambda _: [key for key in _.keys() if key not in ['name']][0]
 
 
-    for need in Needs:
+        for need in Needs:
 
-        if not need.body in ['.:{}\n```{}: {}```'.format(lang, tool(_), _[tool(_)]) for _ in needs]:
-            need.delete()
+            if not need.body in ['.:{}\n```{}: {}```'.format(lang, tool(_), _[tool(_)]) for _ in needs]:
+                need.delete()
